move pollingPlaceRequests to src/actions
